Better way to resize
figure out what size to use so that w*h < maximum
number of pixels, whichever the aspect ratio is

+++ src/surl-server/helpers/stream.c
@@ -706,8 +706,8 @@ next_file:
 
   if (i > FPS && (i % (15*FPS)) == 0) {
     duration = i/FPS;
-    printf("%d seconds, %d frames skipped / %d: %d fps\n", duration,
-         skipped, i, (i-skipped)/duration);
+    printf("%d seconds, %d frames skipped / %d: %.2f fps\n", duration,
+         skipped, i, (float)(i-skipped)/duration);
 
   }
 
@@ -832,8 +832,8 @@ close_last:
   }
   if (i - skipped > FPS) {
     duration = i/FPS;
-    printf("%d seconds, %d frames skipped / %d: %d fps\n", duration,
-          skipped, i, (i-skipped)/duration);
+    printf("%d seconds, %d frames skipped / %d: %.2f fps\n", duration,
+          skipped, i, (float)(i-skipped)/duration);
   }
 
 cleanup:
